Use RadioButton instead of Button
The behavior of this felt more like a ToggleButton or a RadioButton,
where clicking one would change the state and, effectively, turn off the
others.  To make that a bit more apparent in the code:
* Instead of using command objects to change the state, just have
  MainWindowViewModel keep track of its own state.

The code for the IsIn{AddLine,AddPoint,SelectItem}Mode properties looks
like it does because I took a detour through using ToggleButton.  So the
NotifyPropertyChanged calls allow for other `ToggleButton`s (and, really,
any other controls that want to bind to those properties) to be, well,
notified of changes.