It's possible that I've done something truly terrible. I probably should have checked in a few of these changes days ago.
A lot of "refactoring" by that I mean that I moved some things around, built a ton of interfaces, tweaked the class hierarchy and converted the tool functionality to use a strategy pattern.

I plan to write up an explanation of how some of these things work together. Specifically, how the MouseEvent behaviors work together to accomplish something very similar to event routing by using class polymorphism.

Anyway, I've moved the tool selection back into commands because the SetTarget tool has some prerequesites and Commands have built in button state management for this sort of thing.

This could certainly benefit from some testing, but since I'm really bad at writing tests, I'll take any help I can get in that regard.

Anyway, what follows is the original check in comment I made from a few days ago when this changeset was a lot smaller.

So, I've added a tool to set a target location for a shape. At the moment, it only works on points. But the basic idea is sound I think. It will need some serious refactoring before the idea can be broadened to include not only points and lines, but also an object I've yet to write which will ostensibly group a collection of points and lines into a Shape.
For now though, the intended workflow is as follows:
1. I create a "Shape" (atm: place a point)
2. I select the shape I'd like to specify an optimization target for with the select tool (atm: click the select tool, then click the point)
3. Click where I want the target location to be
4. On mouse down, the tool enters a state where mouse moves are used to scale the tolerance for the target.
Onscreen, this is signaled with a scale transform of the target shape.

All of this happens in a really buggy way, so don't be surprised if it takes a few tries for it to work right.

This seems like a pretty clear indication that it's time to refactor the tools into a set of ITools that define their own enabling conditions (for example: The SetTarget Tool requires an item or group of items to be selected before it can be used)
as well as a definition for how to respond to MouseDown, MouseMove, MouseUp, and any modifier keys that alter the normal behavior.