Added a rectangular marquee selection to the selection tool. Clicking on a single item will select that item, and clear any existing selection. Clicking on an empty part of the canvas will clear any existing selections. A MouseDown->MouseMove->MouseUp chain will create a Selection Rectangle in a way that will be truly exciting for anyone who has never used a computer before, but completely mundane and predictable for everyone else, except for the notable fact that it doesn't quite work right. There's a bug that has something to do with the fact that, for whatever reason, the WPF rectangle object doesn't like to let you construct degenerate rectangles where the Max point has lesser values for either or both components than the Min point. So I've got to fix this. Otherwise, the Selection states of the various objects seem to be functioning alright.
I haven't tested the collision detection on the selection rects with the Line object yet. So there's a fair chance it's not yet working the way a reasonable person would expect. Right now, I'm just hit testing the Position property of any selectable shape objects. So there's no intersection testing for shapes that have an area which needs to be considered.